[FEAT]: реализовала отрисовку профиля согласно маршрутам
