+++ src/components/profile-form/profile-form.module.css
@@ -1,0 +1,3 @@
+.buttons-block {
+  align-self: end;
+}
